% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad_fi.R
\name{cad_fi}
\alias{cad_fi}
\title{Registration Data' Download}
\usage{
cad_fi()
}
\value{
a data.frame
}
\description{
This function returns a data.frame containing the latest registration
data of all funds registered in CVM.
}
\examples{


}
