% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_diarios.R
\name{dados_diarios}
\alias{dados_diarios}
\title{Dados Diarios}
\usage{
dados_diarios(cnpj, start, end)
}
\arguments{
\item{cnpj}{a character vector containing the CNPJ of the required funds}

\item{start}{a starting date, minimum value is 2005-01-01}

\item{end}{an end date}
}
\value{
data.frame
}
\description{
This function returns a data.frame of investment funds daily data from CVM
}
