% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_fidc.R
\name{dados_fidc}
\alias{dados_fidc}
\title{Dados FIDC}
\usage{
dados_fidc(cnpj, start, end, table)
}
\arguments{
\item{cnpj}{a character vector containing the CNPJ of the required funds}

\item{start}{a starting date, minimum value is 2005-01-01}

\item{end}{an end date}

\item{table}{a character value containing a FIDC' table number}
}
\value{
data.frame
}
\description{
This function returns a data.frame of FIDCs monthly data from CVM
}
