% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defs_fidcs}
\alias{defs_fidcs}
\title{CVM FIDC's  data variables deffinitions}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{category}{variable name in CVM FIDCs databases}
\item{segment}{category's description}
\item{item}{sub items of categorys when needed}
}
}
\usage{
defs_fidcs
}
\description{
A dataset containing the deffinitions of the variables in FIDC databases
}
\keyword{datasets}
