% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpl_fundos.R
\name{simpl_funds}
\alias{simpl_funds}
\title{Simplify funds names}
\usage{
simpl_funds(names)
}
\arguments{
\item{names}{funds names to simplify}
}
\value{
character vector
}
\description{
Simplify funds names
}
