% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabs_fidcs}
\alias{tabs_fidcs}
\title{CVM FIDC's databases descriptions}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{tabela}{table's id}
\item{conteudo}{database's content}
}
}
\usage{
tabs_fidcs
}
\description{
A dataset containing the description of the 16 available databases of FIDCs
}
\keyword{datasets}
